+++ Simulacion/src/vehiculos/Moto.scala
@@ -5,7 +5,7 @@ import java.awt.Shape
 import movimiento.Velocidad
 import org.jfree.util.ShapeUtilities
 
-class Moto (var placaM:String,velocidad:Velocidad,aceleracion:Int) extends Vehiculo(placaM)(velocidad,aceleracion){
+class Moto (var placaM:String,velocidad:Velocidad,aceleracion:Double) extends Vehiculo(placaM)(velocidad,aceleracion){
   
   val aleatorioM = scala.util.Random
   var placaGenerada:String = ""
